move convertArrayToMap to iterable util
