+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/register/RegisterViewModel.kt
@@ -46,9 +46,9 @@ import org.smartregister.fhircore.engine.rulesengine.RulesExecutor
 import org.smartregister.fhircore.engine.util.DispatcherProvider
 import org.smartregister.fhircore.engine.util.SharedPreferenceKey
 import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
-import org.smartregister.fhircore.engine.util.convertArrayToMap
 import org.smartregister.fhircore.quest.data.register.RegisterPagingSource
 import org.smartregister.fhircore.quest.data.register.model.RegisterPagingSourceState
+import org.smartregister.fhircore.quest.util.convertArrayToMap
 
 @HiltViewModel
 class RegisterViewModel
